Refactor evaluation code
