+++ tests_selector/evaluation/start.py
@@ -1,51 +1,38 @@
+"""This module contains evaluation code for the tool"""
 import os
 import subprocess
-import sys
-import random
-import ntpath
-
 from tests_selector.utils.git import (
-    get_git_repo,
-    file_diff_data_current,
     get_test_lines_and_update_lines,
-    changed_files_branch,
-    file_diff_data_branch,
+    file_diff_data_between_commits,
+    changed_files_between_commits,
+    get_current_head_hash,
 )
-from tests_selector.evaluation.eval_helper import (
-    select_random_file,
-    delete_random_line,
-    capture_specific_exit_code,
-    capture_all_exit_code,
-    print_remove_test_output,
-    install_dependencies,
-)
-
 from tests_selector.utils.db import (
     DatabaseHelper,
-    ResultDatabaseHelper,
     DB_FILE_NAME,
+)
+from tests_selector.evaluation.results_db import (
+    ResultDatabase,
     RESULTS_DB_FILE_NAME,
 )
+from tests_selector.evaluation.eval_helpers import (
+    checkout_remove_branch,
+    delete_random_lines_and_commit,
+    full_diff_between_commits,
+    capture_specific_exit_code,
+    capture_all_exit_code,
+    clear_remove_branch,
+    print_remove_test_output,
+)
 
-PROJECT_FOLDER = sys.argv[1]
 
-
-def random_remove_test(iterations):
-
-    ans = input("init mapping db? [y/n]: ")
-    if ans == "y":
-        subprocess.run(["tests_selector_init"])
-
+def random_remove_test(iterations, deletes_per_iteration, max_wait):
+    """Delete random lines and evaluate tests sets and pytest exitcodes"""
     if not os.path.isfile(DB_FILE_NAME):
-        print("No mapping.db found")
-        exit(1)
+        print("Running mapping database initialization...")
+        subprocess.run(["tests_selector_init"], check=False)
 
-    deletes = 1
-    max_wait = int(input("max wait time for tests? in seconds "))
-    repo = get_git_repo(PROJECT_FOLDER)
-    git_helper = repo.repo.git
-
-    results_db = ResultDatabaseHelper()
+    results_db = ResultDatabase()
     results_db.init_conn()
     results_db.init_results_db()
 
@@ -53,153 +40,59 @@ def random_remove_test(iterations):
     mapping_db.init_conn()
 
     test_suite_size = mapping_db.get_test_suite_size()
-    project_name = ntpath.basename(PROJECT_FOLDER)
-    commithash = repo.get_head().hash
+    project_name = os.getcwd()
+    init_hash = mapping_db.get_last_update_hash()
     db_size = os.path.getsize("./mapping.db")
     project_id = results_db.store_results_project(
-        project_name, commithash, test_suite_size, db_size
+        project_name, init_hash, test_suite_size, db_size
     )
 
-    test_files, src_files = mapping_db.get_testfiles_and_srcfiles()
+    _, src_files = mapping_db.get_testfiles_and_srcfiles()
 
     for i in range(iterations):
-        random_file = select_random_file(src_files)
-        file_id = random_file[0]
-        filename = random_file[1]
 
-        delete_random_line(filename, PROJECT_FOLDER)
+        # Remove random lines
+        checkout_remove_branch()
+        delete_random_lines_and_commit(deletes_per_iteration, src_files)
 
-        diff = file_diff_data_current(filename, PROJECT_FOLDER)
-        test_lines, updates_to_lines = get_test_lines_and_update_lines(diff)
-        tests_line_level = mapping_db.query_tests_srcfile(test_lines, file_id)
-        tests_file_level = mapping_db.query_all_tests_srcfile(file_id)
-
-        if len(tests_line_level) == 0:
-            exitcode_line = 5  # pytest exitcode for no tests collected
-        else:
-            exitcode_line = capture_specific_exit_code(
-                tests_line_level, PROJECT_FOLDER, max_wait
-            )
-        if len(tests_file_level) == 0:
-            exitcode_file = 5
-        else:
-            exitcode_file = capture_specific_exit_code(
-                tests_file_level, PROJECT_FOLDER, max_wait
-            )
-        exitcode_all = capture_all_exit_code(PROJECT_FOLDER, max_wait)
-
-        git_helper.restore(filename)
-
-        results_db.store_results_data(
-            project_id,
-            deletes,
-            exitcode_line,
-            exitcode_file,
-            exitcode_all,
-            len(tests_line_level),
-            len(tests_file_level),
-            diff,
-        )
-
-        print_remove_test_output(
-            i,
-            project_name,
-            commithash,
-            deletes,
-            test_suite_size,
-            len(tests_line_level),
-            len(tests_file_level),
-            exitcode_line,
-            exitcode_file,
-            exitcode_all,
-            RESULTS_DB_FILE_NAME,
-        )
-
-
-def random_remove_test_multiple(iterations):
-    # remove random lines, commit changes and use branch detection
-    ans = input("init mapping db? [y/n]: ")
-    if ans == "y":
-        subprocess.run(["tests_selector_init"])
-
-    if not os.path.isfile(DB_FILE_NAME):
-        print("No mapping.db found")
-        exit(1)
-
-    deletes = int(input("how many random deletes per iteration?: "))
-    max_wait = int(input("max wait time for tests? in seconds "))
-    install_cmd = input(
-        "install command to run tests (new branch transition might break test running, empty if not needed) "
-    )
-    install_required = install_cmd != ""
-    repo = get_git_repo(PROJECT_FOLDER)
-    git_helper = repo.repo.git
-
-    results_db = ResultDatabaseHelper()
-    results_db.init_conn()
-    results_db.init_results_db()
-    mapping_db = DatabaseHelper()
-    mapping_db.init_conn()
-
-    test_suite_size = mapping_db.get_test_suite_size()
-    project_name = ntpath.basename(PROJECT_FOLDER)
-    commithash = repo.get_head().hash
-    db_size = os.path.getsize("./mapping.db")
-    project_id = results_db.store_results_project(
-        project_name, commithash, test_suite_size, db_size
-    )
-
-    test_files, src_files = mapping_db.get_testfiles_and_srcfiles()
-
-    for i in range(iterations):
-        git_helper.checkout("HEAD", b="eval-test-branch")
-        for j in range(deletes):
-            random_file = select_random_file(src_files)
-            file_id = random_file[0]
-            filename = random_file[1]
-
-            delete_random_line(filename, PROJECT_FOLDER)
-            git_helper.add(filename)
-            git_helper.commit("-m", str(j + 1))
-
-        changed_files = changed_files_branch(PROJECT_FOLDER)
+        # Gets tests based on line-level and file-level change
+        current_git_hash = get_current_head_hash()
+        changed_files = changed_files_between_commits(init_hash, current_git_hash)
         tests_line_level = set()
         tests_file_level = set()
-        for f in changed_files:
-            diff = file_diff_data_branch(f, PROJECT_FOLDER)
-            test_lines, updates_to_lines = get_test_lines_and_update_lines(diff)
+        for filename in changed_files:
+            diff = file_diff_data_between_commits(filename, init_hash, current_git_hash)
+            test_lines, _, _ = get_test_lines_and_update_lines(diff)
+            file_id = mapping_db.save_src_file(filename)
             tests_line = mapping_db.query_tests_srcfile(test_lines, file_id)
             tests_file = mapping_db.query_all_tests_srcfile(file_id)
-            for tl in tests_line:
-                tests_line_level.add(tl)
-            for tf in tests_file:
-                tests_file_level.add(tf)
+            for testfunc_line in tests_line:
+                tests_line_level.add(testfunc_line)
+            for testfunc_file in tests_file:
+                tests_file_level.add(testfunc_file)
 
-        full_diff = git_helper.diff("-U0", "master...")
+        # Get full git diff for analysis
+        full_diff = full_diff_between_commits(init_hash, current_git_hash)
 
-        if install_required:
-            install_dependencies(install_cmd)
+        # Pytest exitcodes for running different test sets
+        exitcode_line = (
+            capture_specific_exit_code(list(tests_line_level), max_wait)
+            if tests_line_level
+            else 5
+        )
+        exitcode_file = (
+            capture_specific_exit_code(list(tests_file_level), max_wait)
+            if tests_file_level
+            else 5
+        )
+        exitcode_all = capture_all_exit_code(max_wait)
 
-        if len(tests_line_level) == 0:
-            exitcode_line = 5  # pytest exitcode for no tests collected
-        else:
-            exitcode_line = capture_specific_exit_code(
-                list(tests_line_level), PROJECT_FOLDER, max_wait
-            )
-        if len(tests_file_level) == 0:
-            exitcode_file = 5
-        else:
-            exitcode_file = capture_specific_exit_code(
-                list(tests_file_level), PROJECT_FOLDER, max_wait
-            )
-        exitcode_all = capture_all_exit_code(PROJECT_FOLDER, max_wait)
+        clear_remove_branch()
 
-        git_helper.checkout("master", "--force")
-        git_helper.branch("-D", "eval-test-branch")
-
+        # Store and print data
         results_db.store_results_data(
             project_id,
-            deletes,
+            deletes_per_iteration,
             exitcode_line,
             exitcode_file,
             exitcode_all,
@@ -207,12 +100,11 @@ def random_remove_test_multiple(iterations):
             len(tests_file_level),
             full_diff,
         )
-
         print_remove_test_output(
             i,
             project_name,
-            commithash,
-            deletes,
+            init_hash,
+            deletes_per_iteration,
             test_suite_size,
             len(tests_line_level),
             len(tests_file_level),
@@ -224,23 +116,12 @@ def random_remove_test_multiple(iterations):
 
 
 def main():
-
-    print("Options:")
-    print("Remove random line, run tests and compare results [1]")
-    print("Remove multiple random lines, run tests and compare results [2]")
-    print(
-        "[2] is slower than [1] for 1 line removal because changes are made in a new branch and pip install required for some projects"
-    )
-    ans = input("Choose option 1 or 2: ")
-
-    if ans == "1":
-        ans = input("how many iterations? ")
-        iters = int(ans)
-        random_remove_test(iters)
-    elif ans == "2":
-        ans = input("how many iterations? ")
-        iters = int(ans)
-        random_remove_test_multiple(iters)
+    """Evaluation script entrypoint"""
+    print("RANDOM LINE REMOVE TESTING")
+    iterations = int(input("How many iterations? "))
+    deletes_per_iteration = int(input("How many line removals per iteration? "))
+    max_wait = int(input("Max wait time for test set running (in seconds)? "))
+    random_remove_test(iterations, deletes_per_iteration, max_wait)
 
 
 if __name__ == "__main__":
